fix(gl): split X11 icon code into icon_x11.go
platform_x11.go exceeded the 800-line large-files gate; the new
icon/_NET_WM_ICON and WM_CLASS helpers move to their own file.
